requirements.txt
langchain==0.0.247
openai==0.27.8
PyPDF2==3.0.1
streamlit==1.24.1
oauthlib==3.2.2
python-dotenv